storage.tsを責務ごとに分割
- src/types/font.ts: フォント関連の型・定数・関数
- src/types/premium.ts: プレミアム機能制限の型・定数
- src/types/report.ts: レポート型（WeeklyReport, MonthlyReport）
- src/types/index.ts: 全型ファイルのre-export
- storage.tsからre-exportにより後方互換性を維持

Closes #50
