+++ src/types/index.ts
@@ -1,0 +1,60 @@
+// Re-export all types from individual files
+
+// Font types
+export {
+  type FontFamily,
+  type FontCategory,
+  type FontSize,
+  type FontWeight,
+  type FontSettings,
+  type FontDefinition,
+  FONT_CATEGORIES,
+  getFontDefinition,
+  getFontCategory,
+  getFontFamilyCSS,
+  FONT_FAMILY_MAP,
+  FONT_SIZE_MAP,
+  FONT_WEIGHT_MAP,
+  FONT_FAMILY_NAMES,
+  DEFAULT_FONT_SETTINGS
+} from './font';
+
+// Premium types
+export {
+  type PremiumFeature,
+  type FeatureLimits,
+  FEATURE_LIMITS,
+  FREE_TIER_LIMITS
+} from './premium';
+
+// Report types
+export { type WeeklyReport, type MonthlyReport } from './report';
+
+// Storage types (includes re-exports from font, premium, report for backwards compatibility)
+export {
+  type BlockItem,
+  type Schedule,
+  type DailyStat,
+  type SiteTime,
+  type SiteBlockCount,
+  type DashboardDisplaySettings,
+  type DashboardPreset,
+  type VisionSettings,
+  type SupportedLanguage,
+  type AppSettings,
+  type AnalyticsData,
+  type TrackedSite,
+  type UnblockedSite,
+  type UnblockHistory,
+  type StorageSchema,
+  DEFAULT_SETTINGS,
+  DEFAULT_DISPLAY_SETTINGS,
+  DEFAULT_VISION,
+  getCurrentDisplaySettings,
+  DEFAULT_ANALYTICS,
+  DEFAULT_UNBLOCK_HISTORY,
+  DEFAULT_STORAGE
+} from './storage';
+
+// Message types
+export * from './messages';
