+++ src/types/storage.ts
@@ -1,3 +1,39 @@
+// Storage types and defaults
+
+// Re-export from font.ts for backwards compatibility
+export {
+  type FontFamily,
+  type FontCategory,
+  type FontSize,
+  type FontWeight,
+  type FontSettings,
+  type FontDefinition,
+  FONT_CATEGORIES,
+  getFontDefinition,
+  getFontCategory,
+  getFontFamilyCSS,
+  FONT_FAMILY_MAP,
+  FONT_SIZE_MAP,
+  FONT_WEIGHT_MAP,
+  FONT_FAMILY_NAMES,
+  DEFAULT_FONT_SETTINGS
+} from './font';
+
+// Re-export from premium.ts for backwards compatibility
+export {
+  type PremiumFeature,
+  type FeatureLimits,
+  FEATURE_LIMITS,
+  FREE_TIER_LIMITS
+} from './premium';
+
+// Re-export from report.ts for backwards compatibility
+export { type WeeklyReport, type MonthlyReport } from './report';
+
+// Import types needed for this file
+import type { FontSettings } from './font';
+import { DEFAULT_FONT_SETTINGS } from './font';
+
 // Block list item
 export interface BlockItem {
   id: string;
@@ -59,261 +95,6 @@ export interface DashboardPreset extends DashboardDisplaySettings {
   createdAt: string;
 }
 
-// Font settings for customization
-export type FontFamily =
-  | 'system'
-  // Modern
-  | 'inter'
-  | 'roboto'
-  | 'poppins'
-  | 'lato'
-  | 'opensans'
-  | 'nunito'
-  // Elegant
-  | 'playfair'
-  | 'merriweather'
-  | 'lora'
-  | 'crimsontext'
-  // Impact
-  | 'montserrat'
-  | 'oswald'
-  | 'bebasneue'
-  | 'raleway'
-  // Handwriting
-  | 'dancingscript'
-  | 'caveat'
-  // Japanese
-  | 'notosansjp'
-  | 'notoserifjp'
-  | 'mplusrounded';
-
-export type FontCategory =
-  | 'system'
-  | 'modern'
-  | 'elegant'
-  | 'impact'
-  | 'handwriting'
-  | 'japanese';
-export type FontSize = 'sm' | 'md' | 'lg' | 'xl';
-export type FontWeight = 'normal' | 'medium' | 'semibold' | 'bold';
-
-export interface FontSettings {
-  family: FontFamily;
-  size: FontSize;
-  weight: FontWeight;
-}
-
-// Font category definitions
-export interface FontDefinition {
-  family: FontFamily;
-  name: string;
-  css: string;
-  googleFont?: string; // Google Fonts name for loading
-}
-
-export const FONT_CATEGORIES: Record<
-  FontCategory,
-  { name: string; fonts: FontDefinition[] }
-> = {
-  system: {
-    name: 'System',
-    fonts: [
-      {
-        family: 'system',
-        name: 'System Default',
-        css: 'ui-sans-serif, system-ui, sans-serif'
-      }
-    ]
-  },
-  modern: {
-    name: 'Modern',
-    fonts: [
-      {
-        family: 'inter',
-        name: 'Inter',
-        css: "'Inter', sans-serif",
-        googleFont: 'Inter'
-      },
-      {
-        family: 'roboto',
-        name: 'Roboto',
-        css: "'Roboto', sans-serif",
-        googleFont: 'Roboto'
-      },
-      {
-        family: 'poppins',
-        name: 'Poppins',
-        css: "'Poppins', sans-serif",
-        googleFont: 'Poppins'
-      },
-      {
-        family: 'lato',
-        name: 'Lato',
-        css: "'Lato', sans-serif",
-        googleFont: 'Lato'
-      },
-      {
-        family: 'opensans',
-        name: 'Open Sans',
-        css: "'Open Sans', sans-serif",
-        googleFont: 'Open+Sans'
-      },
-      {
-        family: 'nunito',
-        name: 'Nunito',
-        css: "'Nunito', sans-serif",
-        googleFont: 'Nunito'
-      }
-    ]
-  },
-  elegant: {
-    name: 'Elegant',
-    fonts: [
-      {
-        family: 'playfair',
-        name: 'Playfair Display',
-        css: "'Playfair Display', serif",
-        googleFont: 'Playfair+Display'
-      },
-      {
-        family: 'merriweather',
-        name: 'Merriweather',
-        css: "'Merriweather', serif",
-        googleFont: 'Merriweather'
-      },
-      {
-        family: 'lora',
-        name: 'Lora',
-        css: "'Lora', serif",
-        googleFont: 'Lora'
-      },
-      {
-        family: 'crimsontext',
-        name: 'Crimson Text',
-        css: "'Crimson Text', serif",
-        googleFont: 'Crimson+Text'
-      }
-    ]
-  },
-  impact: {
-    name: 'Impact',
-    fonts: [
-      {
-        family: 'montserrat',
-        name: 'Montserrat',
-        css: "'Montserrat', sans-serif",
-        googleFont: 'Montserrat'
-      },
-      {
-        family: 'oswald',
-        name: 'Oswald',
-        css: "'Oswald', sans-serif",
-        googleFont: 'Oswald'
-      },
-      {
-        family: 'bebasneue',
-        name: 'Bebas Neue',
-        css: "'Bebas Neue', sans-serif",
-        googleFont: 'Bebas+Neue'
-      },
-      {
-        family: 'raleway',
-        name: 'Raleway',
-        css: "'Raleway', sans-serif",
-        googleFont: 'Raleway'
-      }
-    ]
-  },
-  handwriting: {
-    name: 'Handwriting',
-    fonts: [
-      {
-        family: 'dancingscript',
-        name: 'Dancing Script',
-        css: "'Dancing Script', cursive",
-        googleFont: 'Dancing+Script'
-      },
-      {
-        family: 'caveat',
-        name: 'Caveat',
-        css: "'Caveat', cursive",
-        googleFont: 'Caveat'
-      }
-    ]
-  },
-  japanese: {
-    name: 'Japanese',
-    fonts: [
-      {
-        family: 'notosansjp',
-        name: 'Noto Sans JP',
-        css: "'Noto Sans JP', sans-serif",
-        googleFont: 'Noto+Sans+JP'
-      },
-      {
-        family: 'notoserifjp',
-        name: 'Noto Serif JP',
-        css: "'Noto Serif JP', serif",
-        googleFont: 'Noto+Serif+JP'
-      },
-      {
-        family: 'mplusrounded',
-        name: 'M PLUS Rounded 1c',
-        css: "'M PLUS Rounded 1c', sans-serif",
-        googleFont: 'M+PLUS+Rounded+1c'
-      }
-    ]
-  }
-};
-
-// Helper to get font definition by family
-export function getFontDefinition(family: FontFamily): FontDefinition {
-  for (const category of Object.values(FONT_CATEGORIES)) {
-    const font = category.fonts.find((f) => f.family === family);
-    if (font) return font;
-  }
-  return FONT_CATEGORIES.system.fonts[0];
-}
-
-// Helper to get category for a font family
-export function getFontCategory(family: FontFamily): FontCategory {
-  for (const [categoryKey, category] of Object.entries(FONT_CATEGORIES)) {
-    if (category.fonts.some((f) => f.family === family)) {
-      return categoryKey as FontCategory;
-    }
-  }
-  return 'system';
-}
-
-// Weekly report (Premium)
-export interface WeeklyReport {
-  weekStart: string; // YYYY-MM-DD
-  weekEnd: string; // YYYY-MM-DD
-  totalWasteTime: number; // seconds
-  totalInvestTime: number; // seconds
-  totalBlockCount: number;
-  dailyBreakdown: DailyStat[];
-  topWasteSites: { domain: string; time: number }[];
-  topInvestSites: { domain: string; time: number }[];
-  trend: 'improving' | 'declining' | 'stable';
-}
-
-// Monthly report (Premium)
-export interface MonthlyReport {
-  month: string; // YYYY-MM
-  totalWasteTime: number; // seconds
-  totalInvestTime: number; // seconds
-  totalBlockCount: number;
-  weeklyBreakdown: {
-    weekStart: string;
-    wasteTime: number;
-    investTime: number;
-  }[];
-  topWasteSites: { domain: string; time: number }[];
-  topInvestSites: { domain: string; time: number }[];
-  trend: 'improving' | 'declining' | 'stable';
-}
-
 // Vision/Dashboard settings
 export interface VisionSettings {
   // Default display settings (used when no preset is active)
@@ -361,18 +142,6 @@ export interface UnblockHistory {
   sites: Record<string, UnblockedSite>; // key: domain
 }
 
-// Premium feature identifiers
-export type PremiumFeature =
-  | 'unlimited_blocklist'
-  | 'custom_background'
-  | 'dashboard_presets'
-  | 'unsplash'
-  | 'unlimited_history'
-  | 'weekly_report'
-  | 'monthly_report'
-  | 'github_integration'
-  | 'unblock_analytics'; // View unblocked site usage time and re-block
-
 // Complete storage schema
 export interface StorageSchema {
   settings: AppSettings;
@@ -387,12 +156,6 @@ export const DEFAULT_SETTINGS: AppSettings = {
   schedules: [],
   paused: false,
   language: null // Use browser language by default
-};
-
-export const DEFAULT_FONT_SETTINGS: FontSettings = {
-  family: 'system',
-  size: 'md',
-  weight: 'bold'
 };
 
 export const DEFAULT_DISPLAY_SETTINGS: DashboardDisplaySettings = {
@@ -451,65 +214,3 @@ export const DEFAULT_STORAGE: StorageSchema = {
   analytics: DEFAULT_ANALYTICS,
   unblockHistory: DEFAULT_UNBLOCK_HISTORY
 };
-
-// Feature limits type
-export interface FeatureLimits {
-  maxBlockList: number;
-  historyDays: number;
-  maxPresets: number;
-}
-
-// Feature limits
-export const FEATURE_LIMITS: {
-  free: FeatureLimits;
-  premium: FeatureLimits;
-} = {
-  free: {
-    maxBlockList: Infinity, // Unlimited for all users
-    historyDays: 7,
-    maxPresets: 1
-  },
-  premium: {
-    maxBlockList: Infinity,
-    historyDays: Infinity,
-    maxPresets: 5
-  }
-};
-
-// Legacy export for backwards compatibility
-export const FREE_TIER_LIMITS = FEATURE_LIMITS.free;
-
-// Font family CSS mappings (uses new FONT_CATEGORIES)
-export const getFontFamilyCSS = (family: FontFamily): string => {
-  return getFontDefinition(family).css;
-};
-
-// Legacy compatibility - dynamically generated
-export const FONT_FAMILY_MAP: Record<string, string> = Object.values(
-  FONT_CATEGORIES
-)
-  .flatMap((cat) => cat.fonts)
-  .reduce((acc, font) => ({ ...acc, [font.family]: font.css }), {});
-
-// Font size Tailwind class mappings
-export const FONT_SIZE_MAP: Record<FontSize, string> = {
-  sm: 'text-2xl',
-  md: 'text-3xl',
-  lg: 'text-4xl',
-  xl: 'text-5xl'
-};
-
-// Font weight Tailwind class mappings
-export const FONT_WEIGHT_MAP: Record<FontWeight, string> = {
-  normal: 'font-normal',
-  medium: 'font-medium',
-  semibold: 'font-semibold',
-  bold: 'font-bold'
-};
-
-// Font family display names (uses new FONT_CATEGORIES)
-export const FONT_FAMILY_NAMES: Record<string, string> = Object.values(
-  FONT_CATEGORIES
-)
-  .flatMap((cat) => cat.fonts)
-  .reduce((acc, font) => ({ ...acc, [font.family]: font.name }), {});
